wip: dynamic snips
