+++ frontend/docs/components/code/index.tsx
@@ -1,0 +1,4 @@
+export * from './GithubSnippets'
+export * from './CodeRenderer'
+export * from './codeData'
+export * from './codeParser'
